Revert "try to fix send-form"
This reverts commit 094e857c924f8d7f188e2b0472e1b453233ac495.

+++ include/minecraft/net/NetworkIdentifier.h
@@ -11,7 +11,7 @@ public:
     int family;
 
     bool equalsTypeData(NetworkIdentifier const&) const;
-    size_t getHash() const;
+    int getHash() const;
     std::string toString() const;
 
     bool operator==(NetworkIdentifier const& addr) const { return equalsTypeData(addr); }
